Move TgBot to service + fix pannel restart bug #34
